Move onPaymentComplete to useCreatePaymentCompleteCallback
